+++ djangae/fields/counting.py
@@ -1,16 +1,18 @@
 import random
+import warnings
 
 from django.core.exceptions import ImproperlyConfigured
 from django.db import models
 from google.appengine.datastore.datastore_rpc import BaseConnection
 from google.appengine.datastore.datastore_stub_util import _MAX_EG_PER_TXN
 
+from djangae.db import transaction
 from djangae.fields.related import (
     RelatedSetField,
     RelatedIteratorManagerBase,
     ReverseRelatedObjectsDescriptor,
 )
-from djangae.db import transaction
+from djangae.utils import DjangaeDeprecation
 
 
 MAX_ENTITIES_PER_GET = BaseConnection.MAX_GET_KEYS
@@ -26,14 +28,13 @@ class RelatedShardManager(RelatedIteratorManagerBase, models.Manager):
         the additional increment()/decrement()/reset() methods for the counting.
     """
 
-    def __getattr__(self, name):
-        """ Disable count() because it's ambiguous. """
-        if name == 'count':
-            raise AttributeError(
-                "RelatedShardManager has no attribute 'count'. Use 'shard_count()' for the "
-                "number of shards or 'value()' for the total counter value."
-            )
-        return super(RelatedShardManager, self).__getattr__(name)
+    def count(self):
+        warnings.warn(
+            "RelatedShardManager.count is deprecated. Use 'shard_count()' for the "
+            "number of shards or 'value()' for the total counter value.",
+            DjangaeDeprecation, 2
+        )
+        return super(RelatedShardManager, self).count()
 
     def increment(self, step=1):
         if step < 0: